+++ packages/node-core/src/db/migration-service/SchemaMigration.service.ts
@@ -79,6 +79,7 @@ export class SchemaMigrationService {
   }
 
   async run(currentSchema: GraphQLSchema | null, nextSchema: GraphQLSchema, transaction?: Transaction): Promise<void> {
+    logger.info('Running schema migration');
     const schemaDifference = SchemaMigrationService.schemaComparator(currentSchema, nextSchema);
     const {
       addedEnums,